Improvements to the Transaction Table in the Memory Tile
Removed a memory issue, in which Transactions were added to the Transaction Table, but never removed.
Added an element to the transaction to ease the computation, of how many elements are in 1 transaction.
Corrected the setting of the countdown for received CacheRequests from the memory controller.
Created a TODO list for @Mumbi