WiP: Added Debug Information
Added functionality to print message contents during debug sessions of the Memory Tile.
Corrected the scheduling of events (from PostTick to Tick)
Added initializers of fields, to satisfy the linter
Removed old code that was commented for a long time.
Removed the exceptions in Bus and replaced them with assertions.
Found a bug that the bus was not set to idle, when no messages were left in the queue, but kept scheduling pop events instead.