Added LMUL and element type fields to MCPUSetVVL
These fields are required to compute VVL accordingly. Imagine that 16 KB of SP is available. Depending on the type (e.g. 64 bit or 32 bit), a different amount of elements fit into this space. If the 16 KB are for 1 register, LMUL informs the MCPU how many multiples of 16 KB are available for the SP.