Delay Transactions, if the SP is not yet allocated
If a vector load is received by the MemTile and the dedicated space in the SP is not yet allocated, the MemTile returns an SP ALLOC command to the requesting core. This command is acknowledged, once the L2 has evicted dirty pages and has been converted into an SP. In the meantime the load proceeds in parallel. However, we have to delay the return of the Transactions carrying the data, till the  ACK from the ALLOC command has arrived. This commit closes issue #91.