Reschedule the outgoing bus, if the transaction was just peeked at.
This is in continuation of commit #c480e80d60db9bcc8e4dec5399e6b812bb2748f3. For the outgoing bus, it is checked, if the NoC is able to accept a packet. Hence the first transaction in the queue is peeked at. If the NoC can accept the transaction, it is popped from the queue and the bus is rescheduled. However, if the NoC is congested, the bus was not reschedule. This commit fixes the issue.