Allow different mem_op latencies based on the type to be processed
The HBM returns 32 Bytes. If we want to load a cache line, the HBM needs to process 2*32 Byte requests. However, in case of e.g. indexed vector loads, we are not interested in the cache line, but most likely in an 8 Byte load. Hence, only 1 request needs to be processed by the HBM.