trace2gnuplot script added, Logger class updated
The trace2gnuplot script allows to follow a particular address through several stations (interfaces) within Coyote. In contrast to paraver, which observes a particular interface (e.g. MemTile->MC) and check, what address passes through it, this new script allows to follow a particular address and records the interfaces, it passes through.
Coincidentally, the script already discovered a bug in one of the test cases, which has been updated as well.