For any vector mem operation consider the VVL that was set at arrival
There might be a latency between the arrival of a vector memory instruction and the actual time, when the VAG computes the addresses. Meanwhile the VVL could have changed again. That is why this commit stores the VVL setting that was current at the time of arrival of the vector instruction and the VAGs take this value.