Last SP reply  indicates operation completed
There was a bug related to the return of the last SP response to an indexed vector operation. The last packet returned to the SP needs to be marked to indicate that this is the last packet and the memory instruction completed. There was a bug, so that this last packet was not marked accordingly.