Fixed bug that caused a loop back in the Memory Tile Wrapper.
Due to this loop back, packets destined for other Memory Tiles were sent back to the sourcing Memory Tile, which then performed the desired memory operation followed by returning the result.
With this fix, the inter memory tile packets are sent to the destined memory tile, which then replies to the source memory tile once the operation completed.

Closing issue #116