A method controlling packets entering the memory tile
This is to be able to simulate back pressure to the NoC, if the memory tile is unable to process incoming packets.
However, packets loading e.g. the indices which are needed to perform a currently executed vector memory operation are still accepted, since they traverse on a high priority plane of the NoC.

Refer to issue #128